Fixed some warnings on mac
git-svn-id: svn://falconpl.org/falcon/modules/feathers/trunk@204 127a115f-7314-4089-a6a3-bc3257795827

+++ regex/regex_ext.cpp
@@ -728,7 +728,7 @@ void s_expand( RegexCarrier *data, const String &orig, String &expanded )
             }
 
             uint32 val = expanded.getCharAt(pos) - 0x30;
-            if ( data->m_matches > val && val < 10 )
+            if ( ((uint32)data->m_matches) > val && val < 10 )
             {
                // is a valid number?
                expanded.change( pos-1, pos+1, orig.subString( data->m_ovector[val*2], data->m_ovector[val*2+1] ) );
